Signed/unsigned warning fix
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@17014 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/test/test_ncbistr.cpp
@@ -548,7 +548,7 @@ int CTestApplication::Run(void)
     }
 
     {{
-        int i = 0;
+        size_t i = 0;
         ITERATE(list<string>, it, split) {
             assert(i < sizeof(split_result) / sizeof(split_result[0]));
             assert(NStr::Compare(*it, split_result[i++]) == 0);
@@ -849,6 +849,9 @@ int main(int argc, const char* argv[] /*, const char* envp[]*/)
 /*
  * ==========================================================================
  * $Log$
+ * Revision 6.30  2004/08/04 16:52:01  vakatov
+ * Signed/unsigned warning fix
+ *
  * Revision 6.29  2004/07/20 18:47:29  ucko
  * Use the Split test-cases for Tokenize(..., eMergeDelims).
  *
